Adds GetAllRegionNames functionality to OTP controller, IsRewardGranted to UserReward, update user reward functionality and user reward controller and adds data validation to UserRewardChallengeRule object as it can be updated via API endpoint.

Also adds the url for regionreward object.

+++ PLNKTN/PLNKTN/Models/Diet.cs
@@ -2,12 +2,12 @@
 {
     public class Diet
     {
-        public int Plant_based { get; set; }
-        public int Beef { get; set; }
-        public int Pork { get; set; }
-        public int Poultry { get; set; }
-        public int Egg { get; set; }
-        public int Dairy { get; set; }
-        public int Seafood { get; set; }
+        public float Plant_based { get; set; }
+        public float Beef { get; set; }
+        public float Pork { get; set; }
+        public float Poultry { get; set; }
+        public float Egg { get; set; }
+        public float Dairy { get; set; }
+        public float Seafood { get; set; }
     }
 }
